requirements.txt
Django>=4.0
django[argon2]
django[spatialite]  
uvicorn
dj-database-url 
psycopg2-binary   
requests
whitenoise
djangorestframework